Split tck functional tests into their own it/ directory so that they can have different configuration, add the flags to make private and static members injection into a warning instead of an error.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=111028210
