+++ compiler/src/it/tck/src/main/java/test/tck/CarModule.java
@@ -15,20 +15,15 @@
  */
 package test.tck;
 
-import dagger.MembersInjector;
 import dagger.Module;
 import dagger.Provides;
-import org.atinject.tck.auto.Engine;
-import org.atinject.tck.auto.V8Engine;
+import org.atinject.tck.auto.Car;
+import org.atinject.tck.auto.Convertible;
 
 @Module
-public class EngineModule {
+class CarModule {
   @Provides
-  Engine provideEngine(MembersInjector<V8Engine> injector) {
-    // This is provided because V8Engine has no @Inject constructor and Dagger requires an @Inject
-    // constructor, however this is a TCK supplied class that we prefer to leave unmodified.
-    V8Engine engine = new V8Engine();
-    injector.injectMembers(engine);
-    return engine;
+  static Car provideConvertible(Convertible convertible) {
+    return convertible;
   }
 }
